Fix UBSA positive in fineparams constructor

+++ c++/cppdlr/dlr_build.cpp
@@ -32,8 +32,8 @@ namespace cppdlr {
        lambda(lambda),
        p(p),
        nmax((int)ceil(lambda)), // TODO: is this a good choice?
-       npom(max((int)ceil(log(lambda) / log(2.0)), 1)),
-       npt(max((int)ceil(log(lambda) / log(2.0)) - 2, 1)),
+       npom(max(ceil(log(lambda) / log(2.0)), 1.0)),
+       npt(max(ceil(log(lambda) / log(2.0)) - 2, 1.0)),
        nom(2 * p * npom),
        nt(2 * p * npt) {
